6436ab6aee3fa5c036e7bdbd SP1H2T10 - Check header access for interviews
