+++ src/components/HeaderLogged/Manager/index.vue
@@ -10,13 +10,13 @@
           <ul class="flex list-none">
             <li
               class="px-6 py-2 mr-2 font-bold text-white rounded-full cursor-pointer focus:outline-none"
-              @click="() => router.push({ name: 'Interview'})">
+              @click="() => router.push({ name: 'Interviews'})">
               Entrevistas
             </li>
             <li
               class="px-6 py-2 mr-2 font-bold text-white rounded-full cursor-pointer focus:outline-none"
               @click="() => router.push({ name: 'Calendar'})">
-              Calendario
+              Calendário
             </li>
             <li
               class="px-6 py-2 mr-2 font-bold text-white rounded-full cursor-pointer focus:outline-none"
@@ -24,7 +24,7 @@
               Perfil
             </li>
             <li
-              class="px-6 py-2 font-bold bg-white rounded-full cursos-pointer fucus:outline-none"
+              class="px-6 py-2 font-bold bg-white rounded-full cursor-pointer focus:outline-none"
               @click="handleLogout">
               {{ logoutLabel }}
             </li>
