+++ src/components/Interviews/InterviewsLogged/index.vue
@@ -1,0 +1,17 @@
+<template>
+  <interview-logged-manager v-if="roleUser === 'ROLE_MANAGER'"/>
+</template>
+<script>
+import InterviewLoggedManager from "./Manager/index.vue";
+export default {
+  components: {
+    InterviewLoggedManager
+  },
+  setup() {
+    const role = window.localStorage.getItem('roles')
+    return {
+      roleUser: role,
+    }
+  }
+}
+</script>
